lift/index.html

<html>
  <head>
    <link href="../tester/testharness.css" rel="stylesheet">
    <script src="../tester/testharness.js"></script>
  </head>
  <body>
    <style>
      body{
        margin: 0;
      }

      .wrapper{
        max-width: 1024px;
        margin: 0 auto;
        position: relative;
        height: 100%;
      }

      .liftCabine{
        border: 4px solid black;
        width: 150px;
        height: 200px;
        position: absolute;
        /*left: 50%;*/
        transform: translateX(-50%);
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
        bottom: 0;
      }

      .liftCabine1{
        left: 40%;
        background-color: red;
      }

      .liftCabine2{
        left: 60%;
        background-color: green;
      }

      .liftCabine.btn-0{
        bottom: -190px;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }

      .liftCabine.btn-1{
        bottom: 0;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }


      .liftCabine.btn-2{
        bottom: 50%;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }


      .liftCabine.btn-3{
        bottom: calc(100% - 210px);
      }

      .buttons, .buttons2{
        width: 50px;
        position: absolute;
        right: 0;
        top: 50%;
        transform: translateY(-50%);
      }

      .buttons1{
        right: 60px;
      }

      .buttons button, .buttons2 button{
        width: 50px;
        height: 50px;
        margin-bottom: 10px;
      }
    </style>
    <div class='wrapper'>
      <div class='liftCabine liftCabine1'>1</div>
      <div class='liftCabine liftCabine2'>2</div>
      <div class='buttons buttons1'>
        1 lift
        <button class='btn-3'>3</button>
        <button class='btn-2'>2</button>
        <button class='btn-1'>1</button>
        <button class='btn-0'>-1</button>
      </div>
      <div class='buttons buttons2'>
        2 lift
        <button class='btn-3'>3</button>
        <button class='btn-2'>2</button>
        <button class='btn-1'>1</button>
        <button class='btn-0'>-1</button>
      </div>
    </div>

<!--lift programme -->
  <script>
    var buttons1 = document.querySelectorAll('.buttons.buttons1 button');
    var buttons2 = document.querySelectorAll('.buttons.buttons2 button');
    var liftCabine1 = document.querySelector('.liftCabine1');
    var liftCabine2 = document.querySelector('.liftCabine2');

    for (var i = 0; i < buttons1.length; i++){
      buttons1[i].addEventListener('click', function(e){
        liftCabine1.className = '';
        liftCabine1.classList.add(e.target.className, 'liftCabine1', 'liftCabine', 'column1');
      });
    }

    for (var i = 0; i < buttons2.length; i++){
      buttons2[i].addEventListener('click', function(e){
        liftCabine2.className = '';
        liftCabine2.classList.add(e.target.className, 'liftCabine2', 'liftCabine', 'column2');
      });
    }

  </script>

<!--testing-->
  <script>

;(function() {
    //test if massive of two buttons created
    var buttons1 = document.querySelectorAll('.buttons.buttons1 button');
    var buttons2 = document.querySelectorAll('.buttons.buttons2 button');
    var cabine1 = document.querySelector('.liftCabine1');
    var cabine2 = document.querySelector('.liftCabine2');

    var twoBtnArrayTest = async_test('testing if massive of two buttons created');

    window.addEventListener('load', twoBtnArrayTest.step_func(function(){
      var amount = 4;
      assert_equals(buttons1.length, amount);
      assert_equals(buttons2.length, amount);
      this.done();
    }));

    // test column1
    function callGroupTest1(i) {
      var testBtns1 = async_test('if i click on btn ' + i + ' of first column then lift1 should have class column1');

      testBtns1.step(function(){
        var btn = buttons1[i];
        btn.addEventListener('click', testBtns1.step_func(function(e){
          assert_true(cabine1.classList.contains('column1'));
          testBtns1.done();

          // tear-down
          cabine1.classList.remove('column1');
        }));

        btn.dispatchEvent(new Event('click'));
      });
    }

    for (var i = 0; i < buttons1.length; i++) {
      callGroupTest1(i);
    }

    //test column2
    function callGroupTest2(i) {
      var testBtns2 = async_test('if i click on btn ' + i + ' of first column then lift1 should have class column2');

      testBtns2.step(function(){
        var btn = buttons2[i];
        btn.addEventListener('click', testBtns2.step_func(function(e){
          assert_true(cabine2.classList.contains('column2'));
          testBtns2.done();

          // tear-down
          cabine1.classList.remove('column2');
        }));

        btn.dispatchEvent(new Event('click'));
      });
    }

    for (var i = 0; i < buttons1.length; i++) {
      callGroupTest2(i);
    }

    //test if lift1 got class of column1 buttons
    function moveLiftTest(i) {
      var testMoveLift = async_test('if i click on button ' + i + ' my lift should have this -> ' + i + ' class');

      testMoveLift.step(function(){
        var btn = buttons1[i];

        btn.addEventListener('click', testMoveLift.step_func(function(e){
          assert_true(cabine1.classList.contains(btn.className));
          testMoveLift.done();

          //tear-down (чистка)
          cabine1.classList.remove(btn.className);
        }));

        btn.dispatchEvent(new Event('click'));
      })
    }

    for (var i = 0; i < buttons1.length; i++){
      moveLiftTest(i);
    }

    //test if lift2 got class of column2 buttons
    function moveLiftTest2(i) {
      var testMoveLift2 = async_test('if i click on button ' + i + ' my lift should have this -> ' + i + ' class');

      testMoveLift2.step(function(){
        var btn = buttons2[i];

        btn.addEventListener('click', testMoveLift2.step_func(function(e){
          assert_true(cabine2.classList.contains(btn.className));
          testMoveLift2.done();

          //tear-down (чистка)
          cabine2.classList.remove(btn.className);
        }));

        btn.dispatchEvent(new Event('click'));
      })
    }

    for (var i = 0; i < buttons2.length; i++){
      moveLiftTest2(i);
    }

    //test if lift1 got only four classes: liftcabine + liftcabine1 + column1 and class of floor
    function checkClasses(i) {
      var testCheckClasses = async_test('if i click on button btn-' + i + ' elevator1 should have -> btn-' + i + ' class and classes liftcabine and liftcabine1 and column1');

      testCheckClasses.step(function(){
        var btn = buttons1[i];

        btn.addEventListener('click', testCheckClasses.step_func(function(e){
          assert_true(cabine1.classList.contains('liftCabine', 'liftCabine1', e.target.className));
          assert_equals(cabine1.classList.length, 4);
          testCheckClasses.done();

          //tear-down (чистка)
          cabine1.classList.remove(e.target.className);
        }));

        btn.dispatchEvent(new Event('click'));
      });
    }

    for (var i = 0; i < buttons1.length; i++){
      checkClasses(i);
    }

    // test if lift2 got only four classes: liftcabine + liftcabine1 + column2 and class of floor
    function checkClasses2(i) {
      var testCheckClasses2 = async_test('if i click on button btn-' + i + ' elevator1 should have -> btn-' + i + ' class and classes liftcabine and liftcabine1 and column1');

      testCheckClasses2.step(function(){
        var btn = buttons2[i];

        btn.addEventListener('click', testCheckClasses2.step_func(function(e){
          assert_true(cabine2.classList.contains('liftCabine', 'liftCabine1', e.target.className));
          assert_equals(cabine2.classList.length, 4);
          testCheckClasses2.done();

          //tear-down (чистка)
          cabine2.classList.remove(e.target.className);
        }));

        btn.dispatchEvent(new Event('click'));
      });
    }

    for (var i = 0; i < buttons2.length; i++){
      checkClasses2(i);
    }
})();
  </script>
  </body>
</html>
